requirements.txt
black==20.8b1
Django==3.1.3
djangorestframework==3.12.2
djangorestframework_simplejwt==4.5.0
drf-yasg==1.20.0
isort==5.6.4
psycopg2==2.8.6
pylint==2.6.0
